+++ src/main/java/joshie/harvest/api/crops/StateHandlerDefault.java
@@ -40,7 +40,7 @@ public class StateHandlerDefault<C extends Crop> implements IStateHandler<C> {
 
     @Override
     public IBlockState getState(IBlockAccess world, BlockPos pos, PlantSection section, C crop, int stage, boolean withered) {
-        return getState(stage);
+        return getState( Math.min(maximum, stage));
     }
 
     protected IBlockState getState(int stage) {