+++ src/main/java/joshie/harvest/quests/tutorial/QuestCowCare.java
@@ -12,7 +12,7 @@ import joshie.harvest.api.quests.QuestQuestion;
 import joshie.harvest.buildings.HFBuildings;
 import joshie.harvest.core.helpers.InventoryHelper;
 import joshie.harvest.crops.HFCrops;
-import joshie.harvest.quests.TutorialSelection;
+import joshie.harvest.quests.selection.TutorialSelection;
 import joshie.harvest.tools.ToolHelper;
 import joshie.harvest.town.TownHelper;
 import net.minecraft.entity.Entity;